refactor(auth): useMockAuth → useAuth, relocate context to @/lib/auth
Consumer code no longer carries "mock" vocabulary. The provider still
branches on NEXT_PUBLIC_MOCK_API internally; callers just ask for the
current session via useAuth(). Extracted MockAuthToggle to src/mocks/
since it is genuinely dev-only UI.

- move src/mocks/authContext.tsx → src/lib/auth/authContext.tsx
- rename export useMockAuth → useAuth
- extract MockAuthToggle → src/mocks/MockAuthToggle.tsx
- update 5 prod consumers and 3 test files

No behavioral change.
