+++ src/lib/auth/authContext.tsx
@@ -8,7 +8,6 @@ import {
   useState,
   type ReactNode,
 } from "react";
-import { Switch } from "@umichkisa-ds/web";
 
 type AppSession = Session & { token: string };
 
@@ -68,7 +67,6 @@ export function AuthContextProvider({
     }
   }, [mockIsAdmin, IS_MOCK_MODE]);
 
-  // Logging out clears admin state.
   useEffect(() => {
     if (IS_MOCK_MODE && !mockAuthed && mockIsAdmin) {
       setMockIsAdmin(false);
@@ -96,54 +94,13 @@ export function AuthContextProvider({
   return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
 }
 
-export function useMockAuth(): AuthContextValue {
+export function useAuth(): AuthContextValue {
   const ctx = useContext(AuthContext);
   if (!ctx) {
-    throw new Error("useMockAuth must be used within <AuthContextProvider>");
+    throw new Error("useAuth must be used within <AuthContextProvider>");
   }
   return ctx;
 }
 
-export function MockAuthToggle() {
-  const { isMockMode, isAuthenticated, isAdmin, toggle, toggleIsAdmin } =
-    useMockAuth();
-
-  if (!isMockMode) return null;
-
-  return (
-    <div
-      role="group"
-      aria-label="Mock authentication toggle"
-      className="fixed bottom-4 right-4 z-50 flex flex-col gap-2 rounded-lg border border-border-strong bg-surface px-3 py-2 shadow-md"
-    >
-      <div className="flex items-center gap-2">
-        <Switch
-          checked={isAuthenticated}
-          onChange={toggle}
-          aria-label="Toggle mock authentication"
-        />
-        <span className="relative type-body-sm text-foreground">
-          <span aria-hidden="true" className="invisible">
-            Mock: {MOCK_SESSION.user!.email}
-          </span>
-          <span className="absolute inset-0">
-            {isAuthenticated
-              ? `Mock: ${MOCK_SESSION.user!.email}`
-              : "Mock: logged out"}
-          </span>
-        </span>
-      </div>
-      <div className="flex items-center gap-2">
-        <Switch
-          checked={isAdmin}
-          onChange={toggleIsAdmin}
-          disabled={!isAuthenticated}
-          aria-label="Toggle mock admin"
-        />
-        <span className="type-body-sm text-foreground">Mock: admin</span>
-      </div>
-    </div>
-  );
-}
-
-export type { AppSession };
+export { MOCK_SESSION };
+export type { AppSession, AuthContextValue };
