+++ src/lib/next-auth/useAdmin.ts
@@ -2,7 +2,7 @@ import { useSession } from "next-auth/react";
 import { UserSession } from "./types";
 import { useEffect, useState } from "react";
 import { getIsAdmin } from "@/apis/auth/queries";
-import { useMockAuth } from "@/mocks/authContext";
+import { useAuth } from "@/lib/auth/authContext";
 
 // [NOTE]: I think this hook should be moved to shared folder or something
 
@@ -14,7 +14,7 @@ const useAdmin = () => {
   // Mock-mode short-circuit: admin state is driven by MockAuthToggle, not
   // next-auth. Hooks are called unconditionally to satisfy rules-of-hooks;
   // the branch lives in the return.
-  const mock = useMockAuth();
+  const mock = useAuth();
 
   const { data: session, status: sessionStatus } = useSession() as {
     data: UserSession | undefined;
